improve custom SMTP using seprate abstract class
