+++ api/app/Integrations/Handlers/EmailIntegration.php
@@ -7,7 +7,7 @@ use Illuminate\Support\Facades\Log;
 use Illuminate\Support\Facades\Notification;
 use App\Notifications\Forms\FormSubmissionNotification;
 
-class EmailIntegration extends AbstractIntegrationHandler
+class EmailIntegration extends AbstractEmailIntegrationHandler
 {
     public static function getValidationRules(): array
     {
